requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
channels==4.1.0
channels-redis==4.2.0
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
msgpack==1.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
